lib/File/ConfigDir.pm: fix internal function
Even if the issues reported on cpantesters.org were edge-cases,
the test shows up that the test must be aware for MSWin32 and
the _find_common_base_dir can be rewritten in a more re-usable
way.

The fix will not improve the value for File::ConfigDir but now
it can be re-used elsewhere, when it's unsure whether the deepest
path component has to be thrown away in any situation.

Signed-off-by: Jens Rehsack <sno@netbsd.org>